Renamed det_histogram.py file for clarity
