+++ just_bin_it/histograms/histogram_factory.py
@@ -1,9 +1,9 @@
 import logging
 import time
 
-from just_bin_it.histograms.det_histogram import DetHistogram
 from just_bin_it.histograms.histogram1d import Histogram1d
 from just_bin_it.histograms.histogram2d import Histogram2d
+from just_bin_it.histograms.histogram2d_map import DetHistogram
 
 
 def parse_config(configuration, current_time=None):
