+++ tests/test_histogram_factory.py
@@ -1,8 +1,8 @@
 from copy import deepcopy
 
-from just_bin_it.histograms.det_histogram import DetHistogram
 from just_bin_it.histograms.histogram1d import Histogram1d
 from just_bin_it.histograms.histogram2d import Histogram2d
+from just_bin_it.histograms.histogram2d_map import DetHistogram
 from just_bin_it.histograms.histogram_factory import HistogramFactory
 
 CONFIG_1D = [
